\name{REcharts3}
\alias{REcharts3}
\title{a REcharts3 function demo}
\description{
a REcharts3 function demo
}
\usage{
REcharts3(name)
}
\arguments{
  \item{name}{a word}
}
\details{
nothing
}
\value{
no return
}
\references{
nothing
}
\author{
Dan Zhang
}
\note{
nothing
}
\seealso{
nothing
}
\examples{
REcharts3("world")
}
\keyword{ REcharts3 }