\name{layer}
\alias{layer}
\title{a layer function demo}
\description{
a layer function demo
}
\usage{
layer(name)
}
\arguments{
  \item{name}{a word}
}
\details{
nothing
}
\value{
no return
}
\references{
nothing
}
\author{
Channing Wong
}
\note{
nothing
}
\seealso{
nothing
}
\examples{
layer("world")
}
\keyword{ layer }